listening-comp/requirements.txt
# Core dependencies
boto3>=1.34.0
youtube-transcript-api>=0.6.2
chromadb>=0.4.22

# OpenAI and LangChain dependencies
openai==0.28.0
python-dotenv==1.0.1
langchain>=0.1.9
langchain-community>=0.0.24

# Frontend dependencies
streamlit>=1.32.0 